Move room out of util
